+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- New `--prune-until <target>` option to automatically prune the chain up to the given target (height, unix timestamp or YYYY-MM-DD formatted date)
+
+  Require configuring bitcoind with `prune=1` to allow manual pruning via the RPC.
+
 - New `--no-wait-ibd` option to allow importing addresses without waiting for bitcoind to finish syncing first
 
   Useful for tracking wallets during the IBD of a pruned node, so transactions could be indexed before the blocks get pruned.